requirements_pickup.txt
numpy==1.24.3
flask==2.3.3
requests==2.31.0
pytz==2023.3
apscheduler==3.10.4
shapely==2.0.2
sqlalchemy==2.0.23
pymysql==1.1.0
cryptography==41.0.7
python-dotenv==1.0.0
pyjwt==2.8.0
bcrypt==4.1.2